+++ NEWS.md
@@ -23,6 +23,9 @@ CHANGES
   `.rds`, and `.rdata`). Data encryption is based on the AES-GCM algorithm using
   the `openssl::aes_gcm_encrypt()` function (#675).
 
+* `data_tabulate()` gain a `metrics` argument to allow selection of columns to
+  display ("N", "raw", "valid", and "cumulative") (#689, @elinw).
+
 FIXES
 
 * Fix a test due to R-devel change (#677).